Fix a nonproblem
Change the order in the evaluation rules for normal order.
Different orders in the rules lead to different results for the
`path` function but do not affect the normal form of a reduced term.

This seems to be bug in the test suit and not a bug in the
implementation.